+++ lps33hw_STdC/driver/lps33hw_reg.h
@@ -294,7 +294,7 @@ int32_t lps33hw_read_reg(lps33hw_ctx_t *ctx, uint8_t reg, uint8_t* data,
 int32_t lps33hw_write_reg(lps33hw_ctx_t *ctx, uint8_t reg, uint8_t* data,
                           uint16_t len);
 
-extern float_t lps33hw_from_lsb_to_hpa(int16_t lsb);
+extern float_t lps33hw_from_lsb_to_hpa(uint32_t lsb);
 extern float_t lps33hw_from_lsb_to_degc(int16_t lsb);
 
 int32_t lps33hw_autozero_rst_set(lps33hw_ctx_t *ctx, uint8_t val);
